Updating LoginForm_CombinedState.js to use best practice
Important best practice and recommendation: state updates that depend on some previous state should be performed with the help of a function that's passed to the state-updating function.

When receiving a function as a new state value in the state-updating function, React will call that function for you and pass the latest state value to that function.